add ct offset autodetect

+++ Software/src/battery/Shunt.h
@@ -47,7 +47,7 @@ extern const char* name_for_shunt_type(ShuntType type);
 extern ShuntType user_selected_shunt_type;
 
 // Updateable parameters for the Chademo CT Clamp shunt type. Stored in NVM and modifiable via the webserver.
-extern uint16_t ct_clamp_offset_mV;
+extern float ct_clamp_offset_mV;
 extern uint16_t ct_clamp_nominal_voltage_dV;
 extern uint16_t ct_clamp_nominal_current_A;
 enum class adc_attenuation_enum { ADC_0db = 0, ADC_2_5db, ADC_6db, ADC_11db, Highest };
